move data library from core to .table
