--- python/perspective/perspective/core/examples.py
@@ -1,17 +1,0 @@
-# *****************************************************************************
-#
-# Copyright (c) 2019, the Perspective Authors.
-#
-# This file is part of the Perspective library, distributed under the terms of
-# the Apache License 2.0.  The full license can be found in the LICENSE file.
-#
-
-
-def line(dat, col, index='index', key='key'):
-    from .psp import psp
-    return psp(dat,
-               plugin='xy_line',
-               columns=[index, col],
-               rowpivots=[index],
-               columnpivots=[key],
-               aggregates={index: 'last', col: 'last'})
