--- python/perspective/perspective/tests/core/data/test_pa.py
@@ -1,7 +1,0 @@
-# *****************************************************************************
-#
-# Copyright (c) 2019, the Perspective Authors.
-#
-# This file is part of the Perspective library, distributed under the terms of
-# the Apache License 2.0.  The full license can be found in the LICENSE file.
-#
